+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 ### Changes 
+* [FIX] Fix the 'uptime' metric
+* [FEATURE] Add more debug information
 * [FIX] Fix the error that temperature and fan_speed have wrong label
 
 ## 0.3.0 / 2019-06-28